Adds flow to run base tests and conditionally run other tests
1. Adds flow to bail out on setUp if token is absent in environment.
2. Adds conditionals to run secondary tests only if client is
initialized.

+++ tests.py
@@ -18,15 +18,19 @@ class DOClientTest(unittest.TestCase):
     client = None
 
     def setUp(self):
-        self.client = DOClient(self.token)
+        try:
+            self.client = DOClient(self.token)
+        except BaseException:
+            self.assertRaises(APIAuthError)
 
     def test_valid_client(self):
         """Test valid DOClient instance initalization"""
-        self.assertTrue(hasattr(self.client, "filter_droplets") == True)
-        self.assertNotIsInstance(self.client.droplets, NoneType)
-        self.assertIsInstance(self.client.droplets, list)
-        for droplet in self.client.droplets:
-            self.assertIsInstance(droplet, Droplet)
+        if self.client:
+            self.assertTrue(hasattr(self.client, "filter_droplets") == True)
+            self.assertNotIsInstance(self.client.droplets, NoneType)
+            self.assertIsInstance(self.client.droplets, list)
+            for droplet in self.client.droplets:
+                self.assertIsInstance(droplet, Droplet)
 
     def test_invalid_client(self):
         """Test invalid DOClient instance initalization"""
@@ -37,36 +41,38 @@ class DOClientTest(unittest.TestCase):
 
     def test_domain_methods(self):
         """Tests for domain create, list, get, and delete methods"""
-        name = environ.get("test_domain_name")
-        ip_address = environ.get("test_domain_ip")
-        try:
-            create_response = self.client.create_domain(name, ip_address)
-            self.assertIsInstance(create_response, dict)
-            domain = self.client.get_domain(name)
-            self.assertIsInstance(domain, Domain)
-            domains = self.client.get_domains()
-            self.assertIsInstance(domains, list)
-            all_domains = all([isinstance(domain, Domain)
-                               for domain in domains])
-            self.assertTrue(all_domains)
-            del_response = self.client.delete_domain(name)
-            self.assertIsInstance(del_response, dict)
-        except BaseException, error:
-            self.assertIsInstance(error,
-                (APIAuthError, APIError, InvalidArgumentError))
+        if self.client:
+            name = environ.get("test_domain_name")
+            ip_address = environ.get("test_domain_ip")
+            try:
+                create_response = self.client.create_domain(name, ip_address)
+                self.assertIsInstance(create_response, dict)
+                domain = self.client.get_domain(name)
+                self.assertIsInstance(domain, Domain)
+                domains = self.client.get_domains()
+                self.assertIsInstance(domains, list)
+                all_domains = all([isinstance(domain, Domain)
+                                   for domain in domains])
+                self.assertTrue(all_domains)
+                del_response = self.client.delete_domain(name)
+                self.assertIsInstance(del_response, dict)
+            except BaseException, error:
+                self.assertIsInstance(error,
+                    (APIAuthError, APIError, InvalidArgumentError))
 
     def test_droplet_methods(self):
         """Test droplet functionalities"""
-        droplet = self.client.filter_droplets("f")[0]
-        self.assertIsInstance(droplet, Droplet)
-        snapshots = droplet.get_snapshots()
-        self.assertIsInstance(snapshots, list)
-        for snapshot in snapshots:
-            self.assertIsInstance(snapshot, Snapshot)
-        neighbours = droplet.get_neighbours()
-        self.assertIsInstance(neighbours, list)
-        for neighbour in neighbours:
-            self.assertIsInstance(neighbour, Droplet)
+        if self.client:
+            droplet = self.client.filter_droplets("f")[0]
+            self.assertIsInstance(droplet, Droplet)
+            snapshots = droplet.get_snapshots()
+            self.assertIsInstance(snapshots, list)
+            for snapshot in snapshots:
+                self.assertIsInstance(snapshot, Snapshot)
+            neighbours = droplet.get_neighbours()
+            self.assertIsInstance(neighbours, list)
+            for neighbour in neighbours:
+                self.assertIsInstance(neighbour, Droplet)
